Feat/backend/create content

+++ app/models/move.js
@@ -74,14 +74,14 @@ class Move {
         }
     }
 
-    static async delete(moveId) {
+    static async delete(req, moveId) {
 
         try {
             // Select a move 
-            const query = `DELETE FROM "move" WHERE "id"= $1;`; 
+            const query = `DELETE FROM "move" WHERE "id"= $1 AND user_id = $2;`; 
 
             // Delete the move
-            const result = await client.query(query, [moveId]);
+            const result = await client.query(query, [moveId, req.session.user.id]);
 
             //console.log('result :>> ', result);
